+++ timemachine/cpp/CMakeLists.txt
@@ -76,7 +76,7 @@ pybind11_add_module(${LIBRARY_NAME} SHARED NO_EXTRAS
   src/periodic_torsion.cu
   src/chiral_atom_restraint.cu
   src/chiral_bond_restraint.cu
-  src/integrator.cu
+  src/langevin_integrator.cu
   src/context.cu
   src/barostat.cu
   src/rmsd_align.cpp
